Rebased Code and made a PR (#67)
* Updated modifiers

Updated Boss modifiers and added pin modifiers that can be activated from the shop. It can also be enabled in the inspector for testing.

* Updated the Pins

Added Two new pins and updated to the newer models.

* Added materials to the Area

Added material to everything to give it a more in door look.

* Added more prefabs

Added more of the spot light prefabs to sit the world.

* Added more lights

Added more lights to make the lane brighter